Add javadocs, improve naming convention
