+++ .gitignore
@@ -1,3 +1,7 @@
 bin
 tmp
 distribution
+
+# Ignore imports from InteliJ
+*.iml
+.idea
