+++ resources/build.properties
@@ -28,7 +28,7 @@ sketchbook.location=${user.home}/Documents/Processing
 # For OS X,the following path will direct you into Processing's application
 # package, in case you put Processing inside your Applications folder. 
 
-classpath.local.location=${user.home}/Desktop/OpenSimplexNoise-for-Processing/lib
+classpath.local.location=${user.home}/projects/opensource/OpenSimplexNoise-for-Processing/lib
 #classpath.local.location=/Applications/Processing.app/Contents/Java/core/library
